Extend grammar

+++ tests/006_every_datatype.cs
@@ -1,4 +1,4 @@
 int a = 0;
-float b = 0;
+double b = 0.0;
 bool c = true;
 string d = "Lorem ipsum dolor sit amet";
